NT:-314: Add back link to results page
