+++ UI/Pages/SearchResults.cshtml
@@ -5,6 +5,8 @@
     ViewData["Title"] = "Search results";
 }
 
+<govuk-back-link href="/which-subjects?@Model.Data.ToQueryString()" />
+
 <div class="govuk-grid-row">
 	<div class="govuk-grid-column-full">
