+++ UI/cypress/e2e/subjects.js
@@ -1,13 +1,7 @@
 import { Given, When, Then, Step } from "@badeball/cypress-cucumber-preprocessor";
 import { kebabCase, camelCaseKeyStage } from "../support/utils";
+import { allSubjects } from "../support/step_definitions/subjects-page";
         
-const allSubjects = {
-    "Key stage 1": [ "English", "Maths", "Science"],
-    "Key stage 2": [ "English", "Maths", "Science"],
-    "Key stage 3": [ "English", "Humanities", "Maths", "Modern foreign languages", "Science"],
-    "Key stage 4": [ "English", "Humanities", "Maths", "Modern foreign languages", "Science"]
-}
-
 Given("a user has arrived on the 'Which key stages' page", () => {
     Step(this, "a user has arrived on the 'Which key stages' page for postcode 'AB12CD'");
 });
@@ -47,44 +41,12 @@ Then("they will see all the keys stages as options", () => {
     })
 });
 
-Then("they are shown the subjects for {string}", keystage => {
-    
-    const stages = keystage.split(',').map(s => s.trim());
-    stages.forEach(element => {
-
-        const subjects = allSubjects[element]
-
-        cy.get('h2')
-        .contains(`${element} subjects`)
-        .parent()
-        .within(() =>
-        {
-            cy.get('[data-testid="subject-name"]').each((item, index) => {
-                cy.wrap(item).should('contain.text', subjects[index])
-            });
-        });
-
-    });
-});
-
 Then("they are shown all the subjects under all the keys stages", () => {
     Step(this, "they are shown the subjects for 'Key stage 1'")
     Step(this, "they are shown the subjects for 'Key stage 2'")
     Step(this, "they are shown the subjects for 'Key stage 3'")
     Step(this, "they are shown the subjects for 'Key stage 4'")
 })
-
-Then("they will see {string} entered for the postcode", postcode => {
-    cy.get('[data-testid="postcode"]>input').should('contain.value', postcode);
-});
-
-Then("they will see {string} selected", keystages => {
-    const stages = keystages.split(',').map(s => s.trim());
-    stages.forEach(element => {
-        cy.get(`input[id="${kebabCase(element)}"]`).should('be.checked');
-    });
-});
-
 
 Then("the subjects are displayed in alphabetical order", () => {
     Step(this, "they are shown the subjects for 'Key stage 1'")
